backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6

# LangGraph / LangChain（版本需保持一致，否则 checkpoint 会冲突）
langgraph==1.0.10
langgraph-checkpoint==4.0.1
langgraph-checkpoint-sqlite==3.0.3
langchain==1.2.10
langchain-core==1.2.17
langchain-openai==1.1.10

# 数据库
aiosqlite==0.20.0

# 工具
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.5.2
sse-starlette==2.1.3

# 图像处理
Pillow>=10.0.0
httpx>=0.27.0
